// src/validators/schemas/view.schema.ts
// Generated by ts-to-zod
import { z } from 'zod';

import { ViewType } from '../../models/interfaces/view';

export const viewTypeSchema = z.nativeEnum(ViewType);

const objectIdSchema = z.any();

const filterQuerySchema = z.any();

const sortOrderSchema = z.any();

export const viewCriteriaSchema = z.object({
  filter: filterQuerySchema.optional(),
  sort: z.record(sortOrderSchema).optional(),
});

export const viewSchema = z.object({
  id: z.union([z.string(), objectIdSchema]).optional(),
  /**
   * Project name
   * @minLength 1 Name is required
   */
  name: z.string().min(1, 'Name is required'),
  /**
   * View type
   * @default ViewType.CONTRACT
   */
  type: viewTypeSchema.default(ViewType.CONTRACT),
  /**
   * User IDs
   */
  userIds: z.union([z.array(z.string()), z.array(objectIdSchema)]),
  criteria: viewCriteriaSchema,
  createdAt: z.date().optional(),
  updatedAt: z.date().optional(),
  createdBy: z.string().optional(),
  updatedBy: z.string().optional(),
});